MI-94 adding conditional logic to generator to overwrite existing clients
